# backend/api-spec.yaml
openapi: 3.0.0
info:
  title: Clod Pet API
  version: 1.0.0
  description: HTTP API for controlling desktop pets
servers:
  - url: http://localhost:8080
    description: Local development server
paths:
...
  /api/pet/load:
    post:
      summary: Load pet definition
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                pet_path:
                  type: string
      description: Load pet data from `animations.json`, or legacy `animations.xml` if the JSON file is absent.
      responses:
        '200':
          description: Pet definition loaded
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                  payload:
                    type: object
                    properties:
                      title:
                        type: string
                      pet_name:
                        type: string
                      tiles_x:
                        type: integer
                      tiles_y:
                        type: integer
                      png_base64:
                        type: string
                      frame_w:
                        type: integer
                      frame_h:
                        type: integer
                      spawns:
                        type: array
                        items:
                          type: object
                          properties:
                            id:
                              type: integer
                            probability:
                              type: integer
                      anim_count:
                        type: integer
...
  /api/llm/stream:
    post:
      summary: Stream AI response tokens
      description: Server-Sent Events (SSE) endpoint for LLM chat streaming
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                messages:
                  type: array
                  items:
                    type: object
                    properties:
                      role:
                        type: string
                      content:
                        type: string
                stream:
                  type: boolean
      responses:
        '200':
          description: SSE token stream
          content:
            text/event-stream:
              schema:
                type: string
...